requirements.txt
anyio==3.6.2
fastapi==0.87.0
idna==3.4
pydantic==1.10.2
sniffio==1.3.0
starlette==0.21.0
typing_extensions==4.4.0

python-dotenv~=0.21.0
# python-decouple==3.5